requirements.txt
reportlab>=4.0.0
openpyxl>=3.1.0
dbt-core>=1.7.0
dbt-postgres>=1.7.0
astronomer-cosmos>=1.10.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0